refactor(benchmark): split preset code/tests out of oversized main files (#100)
main.go and main_test.go crossed the 500-line coding standard after #100;
extract the preset domain (benchmarkPreset, defaultBenchmarkPresets,
resolveBenchmarkPreset, baselinePresetConfig, parseWorkloads) and its
tests into cmd/benchmark/presets.go and presets_test.go, leaving
describe/run/compare/trend logic in main.go.
